Fix to LabelAnnotator to accept generators; cleanup; working on crowdsourcing tutorial notebook
