+++ snorkel/models/__init__.py
@@ -27,7 +27,7 @@ For example, the correct way to define a new type of Context is:
     SnorkelBase.metadata.create_all(snorkel_engine)
 """
 from .meta import SnorkelBase, SnorkelSession, snorkel_engine, snorkel_postgres
-from .context import Context, Document, Sentence, TemporarySpan, Span
+from .context import Context, Document, Sentence, TemporarySpan, Span, RawText
 from .context import construct_stable_id, split_stable_id
 from .candidate import Candidate, candidate_subclass, Marginal
 from .annotation import Feature, FeatureKey, Label, LabelKey, GoldLabel, GoldLabelKey, StableLabel, Prediction, PredictionKey
